:lock: random userId in token

+++ src/models/ordersModel.js
@@ -1,27 +1,27 @@
 const connection = require('./connection');
 
-const createOrder = async (order, userEmail) => {
+const createOrder = async (order, userId) => {
   const { productId, quantity } = order;
-  const query = 'INSERT INTO orders (productId, quantity, userEmail) VALUES (?,?,?)';
-  const [orderCreated] = await connection.execute(query, [productId, quantity, userEmail]);
+  const query = 'INSERT INTO orders (productId, quantity, userId) VALUES (?,?,?)';
+  const [orderCreated] = await connection.execute(query, [productId, quantity, userId]);
   return orderCreated;
 };
 
-const getOneOrder = async (id) => {
-  const query = 'SELECT orders.orderId, orders.quantity, products.productId, products.productName, products.price FROM orders INNER JOIN products ON products.productId = orders.productId WHERE orderId = ?';
-  const [orderSelected] = await connection.execute(query, [id]);
+const getOneOrder = async (id, userId) => {
+  const query = 'SELECT orders.orderId, orders.quantity, products.productId, products.productName, products.price FROM orders INNER JOIN products ON products.productId = orders.productId WHERE orderId = ? AND userId = ?';
+  const [orderSelected] = await connection.execute(query, [id, userId]);
   return orderSelected;
 };
 
-const deleteOrder = async (id) => {
-  const query = 'DELETE FROM orders WHERE orderId = ?';
-  const [orderDeleted] = await connection.execute(query, [id]);
+const deleteOrder = async (id, userId) => {
+  const query = 'DELETE FROM orders WHERE orderId = ? AND userId = ?';
+  const [orderDeleted] = await connection.execute(query, [id, userId]);
   return orderDeleted;
 };
 
-const userOrder = async (email) => {
-  const query = 'SELECT orders.orderId, orders.quantity, products.productId, products.productName, products.price FROM orders INNER JOIN products ON products.productId = orders.productId WHERE userEmail = ?';
-  const [userOrders] = await connection.execute(query, [email]);
+const userOrder = async (userId) => {
+  const query = 'SELECT orders.orderId, orders.quantity, products.productId, products.productName, products.price FROM orders INNER JOIN products ON products.productId = orders.productId WHERE userId = ?';
+  const [userOrders] = await connection.execute(query, [userId]);
   return userOrders;
 };
